+++ src/voice_among_voices_backend/src/storage/files_and_voices.rs
@@ -1,0 +1,160 @@
+use crate::{
+    generate_angle_file, get_caller_owned_tokens, performance_counter, set_timer,
+    split_into_chunks, AudioSample, ByteBuf, Duration, HttpStreamingResponse,
+    StreamingCallbackHttpResponse, StreamingCallbackToken, StreamingStrategy, ANGLE_FILE_CACHE,
+    AUDIO_PARAMETERS, SAMPLES_MEMORY, SIMULATION_PARAMETERS, STREAMING_CALLBACK,
+    VOICE_NODES_MEMORY, ZERO_DEGREE_FILE_CACHE,
+};
+
+pub fn get_file_for_angle(angle: u64) -> HttpStreamingResponse {
+    let beginning_cost = performance_counter(0);
+
+    let mut result: Vec<u8> = vec![];
+
+    VOICE_NODES_MEMORY.with_borrow(|nodes| {
+        SAMPLES_MEMORY.with_borrow(|samples_map| {
+            result = generate_angle_file(
+                angle as f64,
+                nodes,
+                samples_map,
+                &AUDIO_PARAMETERS,
+                &SIMULATION_PARAMETERS,
+            )
+            .unwrap(); // TODO: error handling
+        });
+    });
+
+    // split into chunks
+    let chunks = split_into_chunks(result, &AUDIO_PARAMETERS);
+
+    ANGLE_FILE_CACHE.with_borrow_mut(|cache| {
+        cache.insert(angle as u32, chunks.clone());
+    });
+
+    set_timer(Duration::from_secs(60), move || {
+        // invalidate after 60s for now
+        ANGLE_FILE_CACHE.with_borrow_mut(|cache| {
+            cache.remove(&(angle as u32));
+        });
+    }); //invalidate after 60 seconds
+
+    let total_chunks = chunks.len() as u32;
+
+    let first_chunk = chunks.get(0).cloned().unwrap_or_default();
+    let token = StreamingCallbackToken {
+        angle: angle as u32,
+        chunk_index: 0,
+        chunks: total_chunks,
+        auth_token: None, // TODO: maybe implement this for security purposes
+    };
+
+    let end_cost = performance_counter(0);
+
+    HttpStreamingResponse {
+        status_code: 200,
+        headers: vec![
+            ("content-type".to_string(), "audio/wav".to_string()),
+            ("x-beginning-cost".to_string(), beginning_cost.to_string()), // Profiling header
+            ("x-end-cost".to_string(), end_cost.to_string()),             // Profiling header
+        ],
+        body: ByteBuf::from(first_chunk),
+        upgrade: None,
+        streaming_strategy: create_strategy(token),
+    }
+}
+
+pub fn get_file_for_zero_angle() -> HttpStreamingResponse {
+    ZERO_DEGREE_FILE_CACHE.with_borrow(|chunks| {
+        let total_chunks = chunks.len() as u32;
+
+        let first_chunk = chunks.get(0).cloned().unwrap_or_default();
+        let token = StreamingCallbackToken {
+            angle: 0 as u32,
+            chunk_index: 0,
+            chunks: total_chunks,
+            auth_token: None, // TODO: maybe implement this for security purposes
+        };
+
+        HttpStreamingResponse {
+            status_code: 200,
+            headers: vec![("content-type".to_string(), "audio/wav".to_string())],
+            body: ByteBuf::from(first_chunk),
+            upgrade: None,
+            streaming_strategy: create_strategy(token),
+        }
+    })
+}
+
+pub fn get_streaming_chunk(token: StreamingCallbackToken) -> StreamingCallbackHttpResponse {
+    let chunks: Vec<Vec<u8>>;
+
+    match token.angle {
+        0 => {
+            chunks = ZERO_DEGREE_FILE_CACHE.with_borrow(|cache| cache.clone());
+        }
+        _ => {
+            chunks = match ANGLE_FILE_CACHE.with_borrow(|cache| cache.get(&token.angle).cloned()) {
+                Some(file_chunks) => file_chunks,
+                None => ic_cdk::trap("Cache out of date, connection too slow"),
+            };
+        }
+    }
+
+    if let Some(token) = token.next() {
+        if let Some(chunk) = chunks.get((token.chunk_index) as usize) {
+            StreamingCallbackHttpResponse {
+                headers: vec![],
+                body: ByteBuf::from(chunk.clone()),
+                token: Some(token),
+            }
+        } else {
+            ic_cdk::trap("Chunk not found");
+        }
+    } else {
+        StreamingCallbackHttpResponse {
+            headers: vec![],
+            body: ByteBuf::new(),
+            token: None,
+        }
+    }
+}
+
+fn create_strategy(token: StreamingCallbackToken) -> Option<StreamingStrategy> {
+    let callback = STREAMING_CALLBACK.clone();
+
+    Some(StreamingStrategy::Callback { token, callback })
+}
+
+pub async fn get_caller_voices() -> Result<Option<Vec<AudioSample>>, String> {
+    let tokens = get_caller_owned_tokens()
+        .await
+        .map(|tokens| {
+            tokens
+                .into_iter()
+                .map(|t| t.try_into())
+                .collect::<Result<Vec<i32>, _>>()
+        })?
+        .map_err(|err| err.to_string())?;
+
+    if tokens.len() == 0 {
+        return Ok(None);
+    };
+
+    let mut response_samples: Vec<AudioSample> = Vec::with_capacity(tokens.len());
+
+    VOICE_NODES_MEMORY.with_borrow(|nodes| {
+        SAMPLES_MEMORY.with_borrow(|samples| {
+            for token in tokens {
+                if let Some(node) = nodes.get(token as u64) {
+                    response_samples.push(
+                        samples
+                            .get(node.sample_id)
+                            .expect("No sample - this shouldn't happen"),
+                    );
+                }
+            }
+        });
+    });
+
+    Ok(Some(response_samples))
+}
